[checkpoint_engine] refactor: move delta_sync core under checkpoint_engine
The delta_sync core (DeltaState / encode / wrapper) was placed under
workers/rollout/ back when delta lived in the SGLang rollout ServerAdapter.
Its only consumer now is DeltaCheckpointEngine, so move it to
verl/checkpoint_engine/delta_sync/ (and the test to tests/checkpoint_engine/)
and switch to relative imports. Fixes the awkward checkpoint_engine ->
workers.rollout dependency direction; no behavior change.
